Move PokemonImage to a new file
