+++ app/src/main/java/com/lucas/pokedexcompose/ui/compose/composables/PokemonItemCard.kt
@@ -1,7 +1,6 @@
 package com.lucas.pokedexcompose.ui.compose.composables
 
 import androidx.compose.foundation.BorderStroke
-import androidx.compose.foundation.Image
 import androidx.compose.foundation.background
 import androidx.compose.foundation.border
 import androidx.compose.foundation.layout.*
@@ -10,14 +9,8 @@ import androidx.compose.material.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.platform.LocalContext
-import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
-import coil.compose.AsyncImagePainter
-import coil.compose.rememberAsyncImagePainter
-import coil.request.ImageRequest
-import com.lucas.pokedexcompose.R
 import com.lucas.pokedexcompose.data.remote.responses.PokemonItem
 import com.lucas.pokedexcompose.ui.theme.PokedexComposeTheme
 import com.lucas.pokedexcompose.ui.theme.PokedexPokemonBackground
@@ -45,43 +38,11 @@ fun PokemonItemCard(
         verticalArrangement = Arrangement.Center
     ) {
         PokemonImage(
-            Modifier
-                .size(120.dp)
+            pokemonId = 10,
+            imageSize = 120
         )
         Text(
             text = pokemonItem.name
-        )
-    }
-}
-
-@Composable
-fun PokemonImage(modifier: Modifier = Modifier) {
-    Box(
-        modifier = modifier,
-        contentAlignment = Alignment.Center
-
-    ) {
-        val pokemonPainter = rememberAsyncImagePainter(
-            model = ImageRequest.Builder(LocalContext.current)
-                .data("https://raw.githubusercontent.com/PokeAPI/sprites/master/sprites/pokemon/132.png")
-                .crossfade(true)
-                .build()
-        )
-
-        if (pokemonPainter.state is AsyncImagePainter.State.Loading ||
-            pokemonPainter.state is AsyncImagePainter.State.Error
-        ) {
-            Image(
-                painter = painterResource(id = R.drawable.gray_pokeball),
-                contentDescription = "",
-                modifier = Modifier.size(40.dp)
-            )
-            Spacer(modifier = Modifier.height(10.dp))
-        }
-        Image(
-            painter = pokemonPainter,
-            modifier = Modifier.fillMaxSize(),
-            contentDescription = ""
         )
     }
 }
